Cachettl.MockWeather (cachettl v0.1.0)

Cachettl.MockWeather provides utility functions that are available for emulating the cache operation as part of the test coverage.

Example

Open two terminals. Lets call them Terminal-1 and Terminal-2 Run the following lines in the cache project directory:

# Terminal-1
/cache$ iex  --sname server@localhost -S mix
#=> Erlang/OTP...

# Terminal-2
/cache$ iex --sname client@localhost -S mix
#=> Erlang/OTP...

# Treminal-2: connect to :server@localhost
iex(client@localhost)1> Node.connect(:server@localhost)
#=> true

# Terminal-1: verify connection
iex(server@localhost)1> Node.list()
#=> [:client@localhost]

# Terminal-2: verify connection
iex(client@localhost)2> Node.list()
#=> [:server@localhost]

# Terminal-1: use Observer to view performance and structures of running processes
iex(server@localhost)2> :observer.start()
#=> ...
#=> :ok

# Terminal-1
iex(server@localhost)3> Cachettl.MockWeather.loop_store()

# Terminal-2
iex(client@localhost)3> Node.spawn(:server@localhost, Cachettl.MockWeather, :loop_get, [])

Expect to also get exceptions like this:

...[warning] Worker-LUX Terminated. Reason: {%RuntimeError{message: "faking an exception"}...

...[error] GenServer {Cachettl.WorkerRegistry, "LUX-worker", :fun_0_arity_processor} terminating
** (RuntimeError) faking an exception

It is a deliberate part of the test to verify how the Supervisor handles child-process failure.

Functions

generate_0_arity_fun(value)

@spec generate_0_arity_fun(any()) :: (() -> {:error, binary()} | {:ok, any()})

generate_0_arity_fun/1 generates a 0-arity function that embeds value which is to be computed at a later time. The 0-arity mimics an expensive data processing that takes a range of 2 - 4 seconds to complete. Result is designed to return {:ok, data} or {:error, reason}, or raise/1 as part of the test.

Note: You are not expected to call this function directly.

loop_get()

@spec loop_get() :: no_return()

loop_get/0 calls Cachettl.get/1 in a recursive loop that mimics infrequent data retrieval. It runs every 6 seconds and takes a random key as an argument on every call.

loop_get_single()

@spec loop_get_single() :: no_return()

loop_get_single/0 same as loop_get/0 but takes the same key as argument on every call.

loop_store()

@spec loop_store() :: no_return()

loop_store/0 calls Cachettl.store/3 in a recursive loop that mimics periodic fast data queries. It runs every second, takes random keys/value that represents mock-weather data by cities, and generated TTL in the range of 30 - 60 seconds as arguments on every call.

single_loop_store()

@spec single_loop_store() :: no_return()

loop_store_single/0 same as loop_store/0 but takes a single key/value as argument on every call.